Refactor Oidc response classes
