+++ sdk/src/main/java/com/microsoft/did/sdk/di/SdkModule.kt
@@ -8,6 +8,10 @@ package com.microsoft.did.sdk.di
 import android.content.Context
 import androidx.room.Room
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import com.microsoft.did.sdk.credential.service.protectors.ExchangeResponseFormatter
+import com.microsoft.did.sdk.credential.service.protectors.IssuanceResponseFormatter
+import com.microsoft.did.sdk.credential.service.protectors.OidcResponseFormatter
+import com.microsoft.did.sdk.credential.service.protectors.PresentationResponseFormatter
 import com.microsoft.did.sdk.credential.service.validators.OidcPresentationRequestValidator
 import com.microsoft.did.sdk.credential.service.validators.PresentationRequestValidator
 import com.microsoft.did.sdk.crypto.CryptoOperations
@@ -31,6 +35,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.converter.scalars.ScalarsConverterFactory
+import javax.inject.Named
 import javax.inject.Singleton
 
 /**
@@ -120,5 +125,26 @@ internal class SdkModule {
     @Singleton
     fun defaultValidator(validator: OidcPresentationRequestValidator): PresentationRequestValidator {
         return validator
+    }
+
+    @Provides
+    @Singleton
+    @Named("issuanceResponseFormatter")
+    fun getIssuanceResponseFormatter(responseFormatter: IssuanceResponseFormatter): OidcResponseFormatter {
+        return responseFormatter
+    }
+
+    @Provides
+    @Singleton
+    @Named("presentationResponseFormatter")
+    fun getPresentationResponseFormatter(responseFormatter: PresentationResponseFormatter): OidcResponseFormatter {
+        return responseFormatter
+    }
+
+    @Provides
+    @Singleton
+    @Named("exchangeResponseFormatter")
+    fun getExchangeResponseFormatter(responseFormatter: ExchangeResponseFormatter): OidcResponseFormatter {
+        return responseFormatter
     }
 }
